Remove most builders
